requirements.txt
fastapi==0.131.0
uvicorn[standard]==0.41.0
# pydantic-settings==2.13.1
python-dotenv==1.0.1
structlog==24.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0